Automatic TOC and site info
Add frontmatter to 2nd level pages to make the navigation order correct in the automatic toc. Add site title and description in the config to show correctly on the page. Add OpenAPI logo and link to Getting started (formerly Start here) on the readme to create front page. Remove original static toc.