WIP: Glossary and guide to referencing
This includes both a defitions page (which could be renamed
to glossary if folks like that better) and the start of a detailed
guide to referencing and related topics.

The referencing guide is more of a WIP than the glossary, and
I'm happy to split the glossary into its own PR if folks like it
enough to approve it.  I just wanted to show it in context as this
is how I wrote it all.

I made the referencing information a section as it is too long
and has too many sub-topics to be on a single page.  I tried to
extract the most common topics into the FAQ so that the basics
can be found quickly, which might lead some folks to read the more
deep-dive pages.

Some pages stop rather abruptly as I was getting bogged down and
figured it would be good to solicit feedback before writing too
much more.  I'm happy to rework any/all of this to suit what
typical users want to know.